+++ packages/xrpl/test/models/oracleSet.test.ts
@@ -167,8 +167,24 @@ describe('OracleSet', function () {
   })
 
   it(`throws w/ invalid AssetPrice of PriceDataSeries`, function () {
-    tx.PriceDataSeries[0].PriceData.AssetPrice = '1234'
-    const errorMessage = 'OracleSet: invalid field AssetPrice'
+    // value cannot be parsed as hexadecimal number
+    tx.PriceDataSeries[0].PriceData.AssetPrice = 'ghij'
+    const errorMessage =
+      'OracleSet: Field AssetPrice must be a valid hex string'
+    assert.throws(() => validateOracleSet(tx), ValidationError, errorMessage)
+    assert.throws(() => validate(tx), ValidationError, errorMessage)
+  })
+
+  it(`verifies valid AssetPrice of PriceDataSeries`, function () {
+    // valid string which can be parsed as hexadecimal number
+    tx.PriceDataSeries[0].PriceData.AssetPrice = 'ab15'
+    assert.doesNotThrow(() => validate(tx))
+  })
+
+  it(`throws w/ invalid AssetPrice type in PriceDataSeries`, function () {
+    tx.PriceDataSeries[0].PriceData.AssetPrice = ['sample', 'invalid', 'type']
+    const errorMessage =
+      'OracleSet: Field AssetPrice must be a string or a number'
     assert.throws(() => validateOracleSet(tx), ValidationError, errorMessage)
     assert.throws(() => validate(tx), ValidationError, errorMessage)
   })