+++ packages/ripple-binary-codec/HISTORY.md
@@ -5,6 +5,9 @@
 ### Added
 * Support for the `Batch` amendment (XLS-56).
 
+### Fixed
+- add `MPTAmount` support in `Issue` (rippled internal type)
+
 ## 2.3.0 (2025-2-13)
 
 ### Added